requirements.txt
selenium==4.15.2
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.4
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
pydantic==2.5.0
webdriver-manager==4.0.1
